\docType{data}
\name{wines}
\alias{wines}
\title{Sensory data from french wines.}
\format{A data frame with 8 rows and 14 variables}
\description{
  Subset data from an experiment conducted on commercial
  ports using four assessors.

  Variables:

  \itemize{ \item A1.Red red \item A1.Brown brown \item
  A1.Soft soft \item A1.Plum plum \item A2.Ruby ruby \item
  A2.Tawny tawny \item A2.Intensity intensity \item A3.Red
  red \item A3.Blue blue \item A3.Brown brown \item
  A3.Intensity intensity \item A4.Depth depth \item
  A4.Fresh fresh \item A4.Brightness brightness }
}
\references{
  Williams A.A., Langron S.P. (1984) The use of free-choice
  profiling for the evaluation of commercial ports.
  \emph{Journal of the Science of Food and Agriculture},
  \bold{68}, No. 5, 558-568.
}
\keyword{datasets}

